feature(token)add parseIntAndCompare to delete method to avoid user to delete other user

+++ app/controllers/core.controller.js
@@ -1,3 +1,6 @@
+import ApiError from '../errors/api.error.js';
+import parseIntAndCompare from '../utils/parseint.compare.js';
+
 export default class Controller {
   static datamapper;
 
@@ -37,8 +40,18 @@ export default class Controller {
     return res.status(200).json(row);
   }
 
-  static async delete({ params }, res, next) {
+  static async delete({ params, user }, res, next) {
     const { id } = params;
+    const { userId } = user;
+    // parsing and comparing the id and userId
+    const isEqual = parseIntAndCompare(id, userId);
+    if (!isEqual) {
+      const err = new ApiError(
+        'Vous n\'avez pas les droits pour accéder à ces informations',
+        { httpStatus: 403 },
+      );
+      return next(err);
+    }
     const deleted = await this.datamapper.delete(id);
     if (!deleted) {
       return next();
